fix: "Cambiar tipo de ID de Integer a Long en ProductoRepository"

+++ src/main/java/com/rodolfogodinez/kinalapp/repository/ProductoRepository.java
@@ -4,14 +4,11 @@ import com.rodolfogodinez.kinalapp.entity.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
 
-    // Buscar cosas por estado
     List<Producto> findByEstado(int estado);
 
-    // Por nobre
     List<Producto> findByNombreProductoContainingIgnoreCase(String nombre);
 
-    // verificar si esta dispnible
     List<Producto> findByStockGreaterThan(Integer stock);
 }
